Fix `J.Unary#getSideEffects()` and `SemanticallyEqual` (#2747)
The implementation of `J.Unary#getSideEffects()` should always return the expression itself as a side effect if it is a modifying unary operator (e.g. postfix `++`).

Also, while implementing a test for this indirectly via `InstanceOfPatternMatchTest` I noticed that `SemanticallyEqual` doesn't allow comparing `J.Parentheses` and `J.ControlParentheses` instances. This is problematic in the case of the `InstanceOfPatternMatch` recipe, as in the `instanceof` it would be the former and in the type cast the latter type.

Further, a few more deficiencies in `SemanticallyEqual` have been fixed.