Longer names for instanceof pattern variables
When the `InstanceOfPatternMatch` recipe is used for `instanceof` expressions in `if` conditions, it makes sense to use longer pattern variable names. So an expression of type `ArrayList` will for example have the pattern variable name `list` in an `if` condition. If some other variable in scope already has that name, `arrayList` will be used instead (and as a fallback `arrayList1` and so on).

Pattern variables in other contexts (e.g. ternary operators) will continue to use short names like `l`.