Mitigate issue caused by `instanceof` pattern support in `JavaTemplate`
When a recipe replaces an expression which is part of an `if` condition which additionally also contains an `instanceof` pattern variable, the generation in `BlockStatementTemplateGenerator` may end up incorrectly generating code for the `if` condition. This is especially problematic when the expression being replaced references any of the pattern variables introduced by the `if` condition.

Mitigates the problem, but does not yet cover the case for when the expression being replaced (or rather its replacement) contains references to pattern variables introduced in the same `if` condition.