Extend SimplifyBooleanExpressionVisitor (#155)
* Extend SimplifyBooleanExpressionVisitor

To reduce duplication. Depends on https://github.com/openrewrite/rewrite/pull/3490

* Add `@Nullable` annotation again

Not sure why it was added in the first place. When can the left or right of a binary be `null` in Kotlin?

* Add `@Nullable` annotation again

Not sure why it was added in the first place. When can the left or right of a binary be `null` in Kotlin?

---------

Co-authored-by: Knut Wannheden <knut@moderne.io>