Simplify by removing error-prone attempt to remove unnecessary parenthesis.
It is hard to know exactly how high up in the tree to go to begin a properly-scoped parenthesis removal, leading to a variety of wrong changes. Even the UnnecessaryParenthesis recipe specialized for this task makes incorrect removals if you start its visiting from the wrong point. UnnecessaryParenthesis recipe comes after this in common static analysis anyway, so the complications are not worth it.