Allow `RemoveRedundantTypeCast` to deal with generics
The `RemoveRedundantTypeCast` recipe will now also remove casts when the types involve generics. Further, when the type cast is the expression of a `return` statement, then the target type will be derived from the corresponding method declaration if any (lambdas are not supported here yet).

Note: This implementation relies on openrewrite/rewrite#3655.