Fix performance issue in `InstanceOfPatternMatch`
The `visitBinary()` in `UseInstanceOfPatternMatching` was visiting the left branch twice. This may not sound that bad, but a `J.Binary` will often be heavily skewed, like for example the tree for `a || b || c || d`:

```
           ||
          /  \
        ||    d
       /  \
     ||    c
    /  \
   a    b
```

By visiting the left branch twice at every level, the `a` leaf will be visited a total of 8 times. This grows exponentially with the tree depth.